Move ftp test code to integration tests
