+++ src/integrationTest/java/uk/gov/hmcts/reform/sendletter/services/FtpUploadTest.java
@@ -2,21 +2,19 @@ package uk.gov.hmcts.reform.sendletter.services;
 
 import com.google.common.io.Files;
 import org.junit.Test;
-import uk.gov.hmcts.reform.sendletter.LocalSftpServer;
 import uk.gov.hmcts.reform.sendletter.helper.FtpHelper;
 import uk.gov.hmcts.reform.sendletter.services.zip.ZippedDoc;
 
 import java.io.File;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.hmcts.reform.sendletter.LocalSftpServer.port;
 
 public class FtpUploadTest {
 
     @Test
     public void connects_to_ftp() throws Exception {
         try (LocalSftpServer server = LocalSftpServer.create()) {
-            FtpHelper.getSuccessfulClient(port).testConnection();
+            FtpHelper.getSuccessfulClient(LocalSftpServer.port).testConnection();
         }
     }
 
@@ -24,7 +22,7 @@ public class FtpUploadTest {
     public void uploads_file() throws Exception {
         ZippedDoc doc = new ZippedDoc("hello.zip", "world".getBytes());
         try (LocalSftpServer server = LocalSftpServer.create()) {
-            FtpClient client = FtpHelper.getSuccessfulClient(port);
+            FtpClient client = FtpHelper.getSuccessfulClient(LocalSftpServer.port);
             client.upload(doc, false);
             File[] files = server.pdfFolder.listFiles();
             assertThat(files.length).isEqualTo(1);
